+++ src/database/queries/applicationForms/selectWithPagination.sql
@@ -1,13 +1,12 @@
 SELECT application_form_to_json(application_forms.*) AS object
 FROM application_forms
-	INNER JOIN opportunities ON application_forms.opportunity_id = opportunities.id
 WHERE
-	has_funder_permission(
+	has_opportunity_permission(
 		:authContextKeycloakUserId,
 		:authContextIsAdministrator,
-		opportunities.funder_short_code,
+		opportunity_id,
 		'view',
-		'funder'
+		'opportunity'
 	)
-ORDER BY application_forms.id
+ORDER BY id
 LIMIT :limit OFFSET :offset;