+++ src/database/queries/bulkUploadTasks/selectById.sql
@@ -7,14 +7,12 @@ SELECT
 FROM bulk_upload_tasks
 	INNER JOIN application_forms
 		ON bulk_upload_tasks.application_form_id = application_forms.id
-	INNER JOIN opportunities
-		ON application_forms.opportunity_id = opportunities.id
 WHERE
 	bulk_upload_tasks.id = :bulkUploadTaskId
-	AND has_funder_permission(
+	AND has_opportunity_permission(
 		:authContextKeycloakUserId,
 		:authContextIsAdministrator,
-		opportunities.funder_short_code,
+		application_forms.opportunity_id,
 		'view',
-		'funder'
+		'opportunity'
 	);